Move Nag component into a folder to allow redux mapping
We'll be pushing the nag visibility to redux in the next commit.
