+++ ui/component/nag/index.js
@@ -1,0 +1,6 @@
+import Nag from './view';
+import { connect } from 'react-redux';
+
+const perform = {};
+
+export default connect(null, perform)(Nag);
