+++ ui/modal/modalPublish/view.jsx
@@ -5,7 +5,7 @@ import { Modal } from 'modal/modal';
 import ClaimPreview from 'component/claimPreview';
 import Button from 'component/button';
 import Card from 'component/common/card';
-import Nag from 'component/common/nag';
+import Nag from 'component/nag';
 
 type Props = {
   closeModal: () => void,
