+++ web/component/nag-data-collection.jsx
@@ -1,6 +1,6 @@
 // @flow
 import React from 'react';
-import Nag from 'component/common/nag';
+import Nag from 'component/nag';
 import I18nMessage from 'component/i18nMessage';
 import Button from 'component/button';
 import { useIsMobile } from 'effects/use-screensize';
